move AES work to encryption crate
